Improve the IU of the plugin search-tables
Now the plugin looks better. The input is like a Kibana's input,
the button of search is a lens of fa icons. When a table is shows,
in the top rigth corner of the div there is a link called "Search",
if this button is clicked, the input will appear. Also, there is
another button, like "Cancel" to hide the input and show the
"Search" link.