Fix bug when click on Edit and metadashboard is not defined
This patch fix the problem when you click on Edit and the
metadashboard is not defined. By default, when the metadashboard is
not defined and you click on Edit, you werent be able to entry into
the Edit mode. Now, with this patch, if the metadashboard is not
defined, you can entry directly.